refactor: move html/js to public and namespace
