--- erpnext/accounts/page/bank_reconciliation_tool/bank_transaction_header.html
@@ -1,27 +1,0 @@
-<div class="transaction-header">
-	<div class="level list-row list-row-head text-muted small">
-		<div class="col-sm-1 ellipsis hidden-xs">
-			{{ __("Date") }}
-		</div>
-		<div class="col-xs-8 col-sm-6 ellipsis list-subject">
-			{{ __("Description") }}
-		</div>
-		<div class="col-sm-1 ellipsis hidden-xs">
-			{{ __("Deposit") }}
-		</div>
-		<div class="col-sm-1 ellipsis hidden-xs">
-			{{ __("Withdrawal") }}
-		</div>
-		<div class="col-sm-2 ellipsis hidden-xs">
-			{{ __("Reference Number") }}
-		</div>
-		<div class="col-sm-1 ellipsis hidden-xs">
-			{{ __("Extra columns") }}
-		</div>
-		<!-- <div class="col-sm-1 ellipsis hidden-xs">
-			{{ __("Currency") }}
-		</div> -->
-		<!-- <div class="col-sm-1 ellipsis">
-		</div> -->
-	</div>
-</div>
